% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ek_GET.R
\name{ek_course_rawdata}
\alias{ek_course_rawdata}
\title{Get course raw data}
\usage{
ek_course_rawdata(projectId, courseId)
}
\arguments{
\item{projectId}{a valid project id}

\item{courseId}{a valid course id}
}
\value{
a data frame
}
\description{
Get course raw data
}
\examples{
\dontrun{
ek_course_rawdata(87123, 4342289)
}
}
