% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ek_GET.R
\name{ek_project_courses}
\alias{ek_project_courses}
\title{Get project courses}
\usage{
ek_project_courses(projectId)
}
\arguments{
\item{projectId}{a valid project id}
}
\value{
a data frame
}
\description{
Get project courses
}
\examples{
\dontrun{
ek_project_courses(48923)
}
}
