% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ek_GET.R
\name{ek_project_nonrespondents}
\alias{ek_project_nonrespondents}
\title{Get project nonrespondents}
\usage{
ek_project_nonrespondents(projectId)
}
\arguments{
\item{projectId}{a valid project id}
}
\value{
a data frame
}
\description{
Get project nonrespondents
}
\examples{
\dontrun{
ek_project_nonrespondents(98233)
}
}
