% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ek_GET.R
\name{ek_project_rawdata}
\alias{ek_project_rawdata}
\title{Get project raw data}
\usage{
ek_project_rawdata(projectId)
}
\arguments{
\item{projectId}{a valid project id}
}
\value{
a data frame
}
\description{
Get project raw data
}
\examples{
\dontrun{
ek_project_rawdata(132784)
}
}
