% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ek_GET.R
\name{ek_project_respondents}
\alias{ek_project_respondents}
\title{Get project respondents}
\usage{
ek_project_respondents(projectId)
}
\arguments{
\item{projectId}{a valid project id}
}
\value{
a data frame
}
\description{
Get project respondents
}
\examples{
\dontrun{
ek_project_respondents(94832)
}
}
