% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ek_GET.R
\name{ek_project_users}
\alias{ek_project_users}
\title{Get project users}
\usage{
ek_project_users(projectId, courseId)
}
\arguments{
\item{projectId}{a valid project id}

\item{userType}{integer, 3 = instructor, 4 = respondent, 6 = TA}
}
\value{
a data frame
}
\description{
Get project users
}
\examples{
\dontrun{
ek_project_users(98423, courseId = 43423)
}
}
