% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ek_GET.R
\name{ek_projects}
\alias{ek_projects}
\title{Get projects}
\usage{
ek_projects(projectType = 1, projectStatus = 4)
}
\arguments{
\item{projectType}{integer, 1 = Course; 2 = General}

\item{projectStatus}{integer, 1 = Not-Deployed; 2 = Deployed-NotStarted; 3 = In-Progress; 4 = Ended}
}
\value{
a data frame
}
\description{
Get projects
}
\examples{
\dontrun{
ek_projects()
}
}
