% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ek_GET.R
\name{ek_response_rates}
\alias{ek_response_rates}
\title{Get response rates}
\usage{
ek_response_rates(projectId)
}
\arguments{
\item{projectId}{a valid project id}
}
\value{
a data frame
}
\description{
Get response rates
}
\examples{
\dontrun{
ek_response_rates(92834)
}
}
