% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ek_set_domain}
\alias{ek_set_domain}
\title{Set your EvaluationKit Domain}
\usage{
ek_set_domain(domain)
}
\arguments{
\item{domain}{Your EvaluationKit domain}
}
\description{
Set your EvaluationKit Domain
}
\examples{
\dontrun{
ek_set_domain("YOUR_DOMAIN")
}
}
