% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ek_set_token}
\alias{ek_set_token}
\title{Set your EvaluationKit API Token}
\usage{
ek_set_token(token)
}
\arguments{
\item{token}{Your EvaluationKit API Token}
}
\description{
Set your EvaluationKit API Token
}
\examples{
\dontrun{
ek_set_token("YOUR_TOKEN")
}
}
