% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ek_GET.R
\name{ek_survey_questions}
\alias{ek_survey_questions}
\title{Get survey questions}
\usage{
ek_survey_questions(surveyId)
}
\arguments{
\item{surveyId}{a valid survey id}
}
\value{
a data frame
}
\description{
Get survey questions
}
\examples{
\dontrun{
ek_survey_questions(surveyId = 132498)
}
}
