% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ek_GET.R
\name{ek_users}
\alias{ek_users}
\title{Get account users}
\usage{
ek_users(username = NULL, userTypes = NULL,
  includeSubaccounts = FALSE)
}
\arguments{
\item{username}{optional username}

\item{userTypes}{CSV of user types. Example: 1,2 returns all EK Administrators and Node Administrators}

\item{includeSubaccounts}{boolean}
}
\value{
a data frame
}
\description{
Get account users
}
\examples{
\dontrun{
ek_users()
}
}
